safari-600.7-branch/LayoutTests/editing/inserting/return-key-span-start.html

<!DOCTYPE html>
<html>
<body>
<div contenteditable id="root"><span id="dupe">hello</span></div>
<script src="../../resources/dump-as-markup.js"></script>
<script>
var sel = window.getSelection();
sel.setPosition(document.getElementById("dupe"), 0);
document.execCommand('InsertParagraph', false, null)
Markup.description("This sets the selection to the start of a span, and hits the enter key.\n"
    + "Expected behavior: when the span is split, its ID should not be duplicated. See bug 62621.");
Markup.dump(root);
</script>
</body>
</html>
