LayoutTests/editing/inserting/return-key-span-start-expected.txt

This sets the selection to the start of a span, and hits the enter key.
Expected behavior: when the span is split, its ID should not be duplicated. See bug 62621.
| <span>
|   id="dupe"
|   <div>
|     <span>
|       <br>
|   "<#selection-caret>hello"
